Added guest, room, and booking menus. Integrated guest menu into admin menu
BUGS:

The admin user menu, both update and delete, fails and show the "message error".
The exception for not finding a guest fails and returns null.
The toString() method of Guest returns null for user and password since we are not assigning them as they don't make sense.
It should display a message for an empty list when there is no content (I have already done it for Guest, it should be replicated for other classes).